// src/main/javascript/src/api/zod.ts
// Generated by ts-to-zod
import {z} from "zod";

export const iconSchema = z.object({
  code: z.string(),
  css: z.string(),
});

export const singleAirportSchema = z.object({
  code: z.string(),
  name: z.string(),
});

export const bookUrlSchema = z.object({
  name: z.string(),
  urls: z.array(z.string()),
});

export const flightSchema = z.object({
  arrivalTime: z.string(),
  cheapSeats: z.string(),
  company: z.string(),
  companyIata: z.string(),
  date: z.string(),
  departureTime: z.string(),
  destinationAirport: singleAirportSchema,
  duration: z.string(),
  price: z.number(),
  sourceAirport: singleAirportSchema,
});

export const tripSchema = z.object({
  bookUrls: z.array(bookUrlSchema),
  hops: z.array(flightSchema),
  lengthOfStay: z.number(),
  totalPrice: z.number(),
});